+++ src/main/java/uol/compass/Programabolsaopenbanking/repositories/ProductRepository.java
@@ -1,7 +1,7 @@
 package uol.compass.Programabolsaopenbanking.repositories;
 
 
-import java.math.BigDecimal;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -16,8 +16,8 @@ import uol.compass.Programabolsaopenbanking.models.ProductModel;
 public interface ProductRepository extends JpaRepository<ProductModel, Id> {
 
 	Optional<ProductModel> findById(long id);	
-	List<ProductModel> findByPriceGreaterThan (BigDecimal priceBigDecimal);
-	List<ProductModel> findByPriceLessThan (BigDecimal priceBigDecimal);
+	List<ProductModel> findByPriceGreaterThan (double priceBigDecimal);
+	List<ProductModel> findByPriceLessThan (double priceBigDecimal);
